requirements.txt
torch
transformers
accelerator
peft
datasets
faiss-cpu
scikit-learn
tqdm
